Refactor Arena
