+++ velox/common/memory/tests/CMakeLists.txt
@@ -21,7 +21,6 @@ add_executable(
   MemoryPoolTest.cpp
   MemoryUsageTest.cpp
   MemoryUsageTrackerTest.cpp
-  ArenaTest.cpp
   MappedMemoryTest.cpp)
 
 target_link_libraries(
