+++ velox/connectors/hive/tests/FileHandleTest.cpp
@@ -20,7 +20,6 @@
 #include "velox/common/caching/SimpleLRUCache.h"
 #include "velox/common/file/File.h"
 #include "velox/common/file/FileSystems.h"
-#include "velox/common/memory/Arena.h"
 #include "velox/exec/tests/utils/TempFilePath.h"
 
 using namespace facebook::velox;
@@ -42,8 +41,8 @@ TEST(FileHandleTest, localFile) {
       std::make_unique<FileHandleGenerator>());
   auto fileHandle = factory.generate(filename);
   ASSERT_EQ(fileHandle->file->size(), 3);
-  Arena arena;
-  ASSERT_EQ(fileHandle->file->pread(0, 3, &arena), "foo");
+  char buffer[3];
+  ASSERT_EQ(fileHandle->file->pread(0, 3, &buffer), "foo");
 
   // Clean up
   remove(filename.c_str());
